Refactor into base class GenericBakedEventBus
