+++ event/src/main/java/net/md_5/bungee/event/asm/CompiledEventBus.java
@@ -1,0 +1,25 @@
+package net.md_5.bungee.event.asm;
+
+import net.md_5.bungee.event.generic.GenericBakedEventBus;
+
+public class CompiledEventBus extends GenericBakedEventBus
+{
+
+    @Override
+    public void bake()
+    {
+        throw new UnsupportedOperationException( "Not supported yet." ); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void bake(Class<?> eventClass)
+    {
+        throw new UnsupportedOperationException( "Not supported yet." ); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void post(Object event)
+    {
+        throw new UnsupportedOperationException( "Not supported yet." ); //To change body of generated methods, choose Tools | Templates.
+    }
+}
